Moved node to models
